requirements.txt
python-telegram-bot>=20.0
aiosqlite
aiohttp
python-dotenv
flask
gunicorn
